line_provider/requirements.txt
fastapi==0.115.6
SQLAlchemy==2.0.37
starlette==0.41.3
pydantic==2.10.5
pydantic-settings==2.7.1
pydantic_core==2.27.2
PyJWT==2.10.1
python-multipart==0.0.20
typing_extensions==4.12.2
annotated-types==0.7.0
aio-pika==9.5.4
aioredis==2.0.1
anyio==4.8.0
sniffio==1.3.1
idna==3.10
uvicorn==0.34.0
click==8.1.8
h11==0.14.0
python-dotenv==1.0.1
email_validator==2.2.0